Two different set of changes.
1.  Progress on the tools to enable coupled Stokes-Darcy.  This includes that ability to "fix" a row in a Jacobian that is all zeros, and access to the boundary tangents in the worksets.

2.  Additional 1D examples.  Turn out, HDIV functionality does not quite work properly in 1D.  Of course, 1D R-T basis functions are the same as the HGRAD basis, so this is what we now use in 1D.  Adding a regression test for 1D mixed Poisson as well as a regression test for the multiscale version.

+++ src/subgrid/subgridModel.hpp
@@ -122,7 +122,7 @@ namespace MrHyDE {
     Kokkos::View<int*,AssemblyDevice> macro_numDOF;
     
     vector<string> macro_paramnames, macro_disc_paramnames, macrosidenames;
-    int macro_block;
+    size_t macro_block;
     ScalarT cost_estimate;
     bool subgrid_static;
